chore: create new module graph_files
