+++ ra2ce/common/configuration/config_wrapper_protocol.py
@@ -26,7 +26,7 @@ from pathlib import Path
 from typing import Optional, Protocol, runtime_checkable
 
 from ra2ce.common.configuration.config_data_protocol import ConfigDataProtocol
-from ra2ce.graph.graph_files_collection import GraphFilesCollection
+from ra2ce.graph.graph_files.graph_files_collection import GraphFilesCollection
 
 
 @runtime_checkable
