+++ ra2ce/graph/graph_files/graph_files_enum.py
@@ -1,0 +1,10 @@
+from enum import Enum
+
+
+class GraphFileEnum(Enum):
+    BASE_GRAPH = 1
+    BASE_GRAPH_HAZARD = 2
+    ORIGINS_DESTINATIONS_GRAPH = 3
+    ORIGINS_DESTINATIONS_GRAPH_HAZARD = 4
+    BASE_NETWORK = 5
+    BASE_NETWORK_HAZARD = 6
