+++ tests/analyses/indirect/test_origin_closest_destination.py
@@ -3,7 +3,7 @@ from ra2ce.analyses.analysis_config_data.analysis_config_data import (
     AnalysisSectionIndirect,
 )
 from ra2ce.analyses.indirect.origin_closest_destination import OriginClosestDestination
-from ra2ce.graph.graph_files_collection import GraphFilesCollection
+from ra2ce.graph.graph_files.graph_files_collection import GraphFilesCollection
 from ra2ce.graph.network_config_data.network_config_data import (
     NetworkSection,
     OriginsDestinationsSection,
