+++ tests/graph/test_network_config_wrapper.py
@@ -2,7 +2,7 @@ import shutil
 
 import pytest
 
-from ra2ce.graph.graph_files_collection import GraphFilesCollection
+from ra2ce.graph.graph_files.graph_files_collection import GraphFilesCollection
 from ra2ce.graph.network_config_wrapper import NetworkConfigWrapper
 from tests import test_results
